Rework Registry & Branch to remove circular imports
